move visualizer controls to widgets.py
